# swagger/swagger.json
{
  "swagger": "2.0" ,
  "info": {
    "version": "1.0.0" ,
    "title": "Micro Credits Backend" ,
    "description": "Micro Credits Backend" ,
    "license": {
      "name": "MIT" ,
      "url": "https://opensource.org/licenses/MIT"
    }
  } ,
  "host": "ec2-54-171-157-12.eu-west-1.compute.amazonaws.com:3005",
...
  "schemes": [
    "http"
  ] ,
  "consumes": [
    "application/json"
  ] ,
  "produces": [
    "application/json"
  ] ,
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security":[
    {
      "api_key":[]
    }
  ],
  "paths":{
...
    "/customers/customerSignUp": {
      "post": {
        "tags": ["Customers"],
        "summary": "Register new customer in the system",
        "description": "Register new customer in the system",
        "parameters": [
          {
            "name": "Customer",
            "in": "body",
            "description":"Data of customer",
            "schema": {
              "$ref": "#/definitions/customerData"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New customer is created" ,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
    "/customers/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of customer record that we want",
          "type": "string"
        }
      ],
      "get": {
        "summary": "Get customer record with give ID",
        "tags": [
          "Customers"
        ],
        "responses": {
          "200": {
            "description": "Customer record fetched",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "summary": "Update customer with give ID",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "Customer with new values of properties",
            "schema": {
              "$ref": "#/definitions/updateUserData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer is updated",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/customers/login": {
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Login route for customer",
        "description": "Allows user to login by providing credentials",
        "parameters": [
          {
            "name": "login",
            "in": "body",
            "description": "login credentials of customer",
            "schema": {
              "$ref": "#/definitions/customerLoginData"
            }
          }
        ],
        "responses" :{
          "200": {
            "description": "Customer logged in successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/customers/otp": {
      "post": {
        "tags": ["Customers"],
        "summary": "Generate Otp and send to customer",
        "parameters": [
          {
            "name": "Otp data",
            "in": "body",
            "description": "Send otp to customer",
            "schema": {
              "$ref": "#/definitions/OtpData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
    "/customers/checkforotp/{id}/{otp}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of customer that we want to find",
          "type": "string"
        },
        {
          "name":"otp",
          "in": "path",
          "required": true,
          "description": "Otp",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "verified Otp",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
...
    "/customerBalance": {
      "get":{
        "tags": [
          "CustomerBalance"
        ],
        "summary": "Gets balances of all customer",
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      },
      "post":{
        "tags": [
          "CustomerBalance"
        ],
        "parameters": [
          {
            "name": "CustomerBalance",
            "in": "body",
            "description": "Body of customer balance that we want to create",
            "schema": {
              "$ref": "#/definitions/customerBalanceData"
            }
          }
        ],
        "summary": "Creates new balance account for customer",
        "responses": {
          "200": {
            "description": "New city is created" ,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/customerBalance/{customerId}": {
      "parameters": [
        {
          "name": "customerId",
          "in": "path",
          "required": true,
          "description": "ID of customer",
          "type": "string"
        }
      ],
      "put": {
        "summary": "Update balance with give customer id",
        "tags": [
          "CustomerBalance"
        ],
        "parameters": [
          {
            "name": "country",
            "in": "body",
            "description": "Country with new values of properties",
            "schema": {
              "$ref": "#/definitions/balanceUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer balance is updated",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "get": {
        "tags": [
          "CustomerBalance"
        ],
        "summary": "Get balance by id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "customerData": {
      "required": [
        "firstName",
        "lastName",
        "streetNo",
        "cityId",
        "stateId",
        "postalCode",
        "countryId",
        "dateOfBirth",
        "countryCode",
        "phoneNumber",
        "email",
        "password",
        "type"
      ],
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "streetNo": {
          "type": "string"
        },
        "cityId": {
          "type": "integer"
        },
        "stateId": {
          "type": "integer"
        },
        "postalCode": {
          "type": "string"
        },
        "countryId": {
          "type": "integer"
        },
        "dateOfBirth": {
          "format":"date",
          "type": "string"
        },
        "countryCode": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "pattern": "^(([^<>()\\[\\]\\.,;:\\s@\"]+(\\.[^<>()\\[\\]\\.,;:\\s@\"]+)*)|(\".+\"))@((\\[[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}])|(([a-zA-Z\\-0-9]+\\.)+[a-zA-Z]{2,}))$"
        },
        "password": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "enum":["phoneNumber", "email"]
        }
      }
    },
    "OtpData": {
      "required": [
        "id",
        "type"
      ],
      "properties": {
        "id": {
          "type": "integer"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "customerLoginData": {
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "updateUserData": {
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "streetNo": {
          "type": "string"
        },
        "cityId": {
          "type": "integer"
        },
        "stateId": {
          "type": "integer"
        },
        "postalCode": {
          "type": "string"
        },
        "countryId": {
          "type": "integer"
        },
        "dateOfBirth": {
          "format":"date",
          "type": "string"
        },
        "countryCode": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
    "balanceUpdate": {
      "required": [
        "balance"
      ],
      "properties": {
        "balance": {
          "type": "number"
        }
      }
    },
    "customerBalanceData":{
      "required": [
        "customerId",
        "balance"
      ],
      "properties": {
        "customerId":{
          "type":"integer"
        },
        "balance": {
          "type": "number"
        }
      }
    }
  }
}